pyproject.toml
# 参考: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#a-full-example

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "problem-awesome-problem" # MUST edit here to problem / repository name
version = "0.0.1"
dependencies = [
    "jsonschema",
    "click",
    "PyYAML",
] # MUST add dependencies for your program
requires-python = ">=3.11" # SHOULD edit required python version for your program
authors = [
    {name = "Hoge", email = "hogehoge@email.com"},
    {name = "Fuga", email = "fugafuga@univ.ac.jp"},
] # SHOULD edit your author information
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
Homepage = "https://ec-comp.jpnsec.org/problems/awesome-problem" # SHOULD edit problem wab page URL in OptHub
Repository = "https://github.com/opthub-org/problem-awesome-problem" # SHOULD edit repository URL

# 開発用の設定
[project.optional-dependencies]
dev = [
    "pytest",
    "pyproject-flake8",
    "mypy >= 0.900",
    "black",
    "isort"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 100
exclude = [".venv", "tests", "build"]

[tool.mypy]
exclude = [".venv", "tests", "build"]
